# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "DIMO Device Data API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/v1/autopi/last-seen/{ethAddr}": {
            "get": {
                "security": [
                    {
                        "PreSharedKey": []
                    }
                ],
                "description": "Specific for AutoPi - get when a device last sent data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "autopi"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "device ethereum address",
                        "name": "ethAddr",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "invalid eth addr"
                    },
                    "404": {
                        "description": "no device found with eth addr or no data found"
                    },
                    "500": {
                        "description": "no device found or no data found, or other transient error"
                    }
                }
            }
        },
        "/v1/user/device-data/{userDeviceID}/daily-distance": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[🔴__Warning - API Shutdown by June 30, 2024, Use `/v2/vehicles/:tokenId/analytics/daily-distance` instead__🔴] Get kilometers driven for a userDeviceID each day.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-data [End Of Life Warning]"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user device id",
                        "name": "userDeviceID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "IANAS time zone id, e.g., America/Los_Angeles",
                        "name": "time_zone",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.DailyDistanceResp"
                        }
                    },
                    "404": {
                        "description": "no device found for user with provided parameters"
                    }
                }
            }
        },
        "/v1/user/device-data/{userDeviceID}/distance-driven": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get kilometers driven for a userDeviceID since connected (ie. since we have data available)\n[🔴__Warning - API Shutdown by June 30, 2024, Use `/v2/vehicles/:tokenId/analytics/total-distance` instead__🔴]  if it returns 0 for distanceDriven it means we have no odometer data.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-data [End Of Life Warning]"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user device id",
                        "name": "userDeviceID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "no device found for user with provided parameters"
                    }
                }
            }
        },
        "/v1/user/device-data/{userDeviceID}/export/json/email": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Enqueues a data export job for the specified device. A link to download\na large JSON file of signals will be emailed to the address on file for the\ncurrent user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-data"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "example": "2OQjmqUt9dguQbJt1WImuVfje3W",
                        "description": "Device id",
                        "name": "userDeviceID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/user/device-data/{userDeviceID}/historical": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all historical data for a userDeviceID, within start and end range",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-data"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "userDeviceID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "startDate eg 2022-01-02. if empty two weeks back",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "endDate eg 2022-03-01. if empty today",
                        "name": "endDate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/user/device-data/{userDeviceID}/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the latest status update for the device. May return 404 if the\nuser does not have a device with the ID, or if no status updates have come. Note this endpoint also exists under nft_controllers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-data"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user device ID",
                        "name": "user_device_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DIMO-Network_device-data-api_internal_response.DeviceSnapshot"
                        }
                    }
                }
            }
        },
        "/v1/vehicle/{tokenID}/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all historical data for a tokenID, within start and end range",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-data"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "token id",
                        "name": "tokenID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "startDate ex: 2022-01-02; or,  2022-01-02T09:00:00Z; if empty two weeks back",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "endDate ex: 2022-03-01; or, 2023-03-01T09:00:00Z; if empty today",
                        "name": "endDate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/vehicle/{tokenId}/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the latest status update for the vehicle with a given token id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-data"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "token id",
                        "name": "tokenId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DIMO-Network_device-data-api_internal_response.DeviceSnapshot"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/v1/vehicle/{tokenId}/status-raw": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the latest status update for the vehicle with a given token id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-data"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "token id",
                        "name": "tokenId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v2/vehicle/{tokenID}/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all historical data for a tokenID, within start and end range",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-data"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "token id",
                        "name": "tokenID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "startTime is an RFC3339 formatted date-time string. If empty two weeks from endTime",
                        "name": "startTime",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "endTime is an RFC3339 formatted date-time string. If empty the current time",
                        "name": "endTime",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "number of data points to return, default 1000",
                        "name": "buckets",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v2/vehicle/{tokenId}/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the latest status update for the vehicle with a given token id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-data"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "token id",
                        "name": "tokenId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_DIMO-Network_device-data-api_internal_response.Device"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/v2/vehicles/{tokenID}/analytics/daily-distance": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get kilometers driven for a userDeviceID each day.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-data"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user device id",
                        "name": "userDeviceID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "IANAS time zone id, e.g., America/Los_Angeles",
                        "name": "time_zone",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.DailyDistanceResp"
                        }
                    },
                    "404": {
                        "description": "no device found for user with provided parameters"
                    }
                }
            }
        },
        "/v2/vehicles/{tokenID}/analytics/total-distance": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get kilometers driven for a userDeviceID since connected (ie. since we have data available)\nif it returns 0 for distanceDriven it means we have no odometer data.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-data"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user device id",
                        "name": "userDeviceID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "no device found for user with provided parameters"
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_DIMO-Network_device-data-api_internal_response.Device": {
            "type": "object",
            "properties": {
                "recordCreatedAt": {
                    "type": "string"
                },
                "recordUpdatedAt": {
                    "description": "Other",
                    "type": "string"
                },
                "status": {
                    "description": "VSS status data",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_DIMO-Network_device-data-api_internal_response.Status"
                        }
                    ]
                }
            }
        },
        "github_com_DIMO-Network_device-data-api_internal_response.DeviceSnapshot": {
            "type": "object",
            "properties": {
                "ambientTemp": {
                    "type": "number"
                },
                "batteryCapacity": {
                    "type": "integer"
                },
                "batteryVoltage": {
                    "type": "number"
                },
                "chargeLimit": {
                    "type": "number"
                },
                "charging": {
                    "type": "boolean"
                },
                "fuelPercentRemaining": {
                    "type": "number"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "odometer": {
                    "type": "number"
                },
                "oil": {
                    "type": "number"
                },
                "range": {
                    "type": "number"
                },
                "recordCreatedAt": {
                    "type": "string"
                },
                "recordUpdatedAt": {
                    "type": "string"
                },
                "soc": {
                    "type": "number"
                },
                "tirePressure": {
                    "$ref": "#/definitions/smartcar.TirePressure"
                }
            }
        },
...
        "internal_controllers.DailyDistanceDay": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "distance": {
                    "type": "number"
                }
            }
        },
        "internal_controllers.DailyDistanceResp": {
            "type": "object",
            "properties": {
                "days": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_controllers.DailyDistanceDay"
                    }
                }
            }
        },
...
        "smartcar.TirePressure": {
            "type": "object",
            "properties": {
                "age": {
                    "description": "Deprecated: Should use DataAge instead of Age",
                    "type": "string"
                },
                "backLeft": {
                    "type": "number"
                },
                "backRight": {
                    "type": "number"
                },
                "dataAge": {
                    "type": "string"
                },
                "frontLeft": {
                    "type": "number"
                },
                "frontRight": {
                    "type": "number"
                },
                "requestId": {
                    "type": "string"
                },
                "unitSystem": {
                    "$ref": "#/definitions/smartcar.UnitSystem"
                }
            }
        },
        "smartcar.UnitSystem": {
            "type": "string",
            "enum": [
                "metric",
                "imperial"
            ],
            "x-enum-varnames": [
                "Metric",
                "Imperial"
            ]
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}